* eos stat/achievement tidy up, stats trigger achievements * fix some formatting

+++ src/main.hpp
@@ -649,7 +649,7 @@ extern std::unordered_map<std::string, std::string> achievementDesc;
 extern std::unordered_set<std::string> achievementHidden;
 typedef std::function<bool(std::pair<std::string, std::string>, std::pair<std::string, std::string>)> Comparator;
 extern std::set<std::pair<std::string, std::string>, Comparator> achievementNamesSorted;
-extern std::unordered_map<std::string, double> achievementProgress;
+extern std::unordered_map<std::string, int> achievementProgress;
 extern std::unordered_map<std::string, int64_t> achievementUnlockTime;
 extern voxel_t** models;
 extern polymodel_t* polymodels;
